requirements.txt
cn_clip
pathlib
math
pandas
gradio
Pillow